# backend/openapi/horse-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Horse Reservation API",
    "description": "마장 예약 서비스 API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/admin/recurring-holidays/{holidayId}": {
      "put": {
        "tags": [
          "admin-recurring-holiday-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "holidayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringHolidayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringHolidayMutationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/schedule-templates/preview": {
      "post": {
        "tags": [
          "admin-schedule-template-controller"
        ],
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleTemplatePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleTemplateImpactResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/recurring-holidays": {
      "get": {
        "tags": [
          "admin-recurring-holiday-controller"
        ],
        "operationId": "getHolidays",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecurringHolidayResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-recurring-holiday-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringHolidayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringHolidayMutationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/recurring-holidays/preview": {
      "post": {
        "tags": [
          "admin-recurring-holiday-controller"
        ],
        "operationId": "preview_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringHolidayPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringHolidayImpactResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/schedule-templates/{templateId}/activation": {
      "patch": {
        "tags": [
          "admin-schedule-template-controller"
        ],
        "operationId": "changeActivation",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleActivationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleTemplateMutationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/recurring-holidays/{holidayId}/activation": {
      "patch": {
        "tags": [
          "admin-recurring-holiday-controller"
        ],
        "operationId": "changeActivation_1",
        "parameters": [
          {
            "name": "holidayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleActivationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringHolidayMutationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/reservations/{reservationId}/cancellation-preview": {
      "get": {
        "tags": [
          "member-reservation-cancel-controller"
        ],
        "operationId": "preview_2",
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCancellationPreviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/schedule-sync": {
      "get": {
        "tags": [
          "admin-schedule-synchronization-controller"
        ],
        "operationId": "getStatus",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleSynchronizationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/reservations/{reservationId}/cancellation-preview": {
      "get": {
        "tags": [
          "admin-reservation-cancel-controller"
        ],
        "operationId": "preview_3",
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "responsibility",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCancellationPreviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ScheduleTemplateImpactResponse": {
        "type": "object",
        "properties": {
          "affectedDateCount": {
            "type": "integer",
            "format": "int64"
          },
          "existingTimeSlotCount": {
            "type": "integer",
            "format": "int64"
          },
          "activeReservationCount": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "activeReservationCount",
          "affectedDateCount",
          "existingTimeSlotCount"
        ]
      },
      "ScheduleTemplateMutationResponse": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/ScheduleTemplateResponse"
          },
          "pendingConfigVersion": {
            "type": "integer",
            "format": "int64"
          },
          "impact": {
            "$ref": "#/components/schemas/ScheduleTemplateImpactResponse"
          },
          "synchronization": {
            "$ref": "#/components/schemas/ScheduleSynchronizationResponse"
          }
        },
        "required": [
          "impact",
          "pendingConfigVersion",
          "synchronization",
          "template"
        ]
      },
...
      "RecurringHolidayRequest": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ]
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          },
          "holidayReason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          },
          "expectedConfigVersion": {
            "type": "integer",
            "format": "int64"
          },
          "changeReason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          }
        },
        "required": [
          "changeReason",
          "dayOfWeek",
          "effectiveFrom",
          "expectedConfigVersion",
          "holidayReason"
        ]
      },
...
      "RecurringHolidayImpactResponse": {
        "type": "object",
        "properties": {
          "previous": {
            "$ref": "#/components/schemas/RecurringHolidayImpactCountsResponse"
          },
          "current": {
            "$ref": "#/components/schemas/RecurringHolidayImpactCountsResponse"
          },
          "combined": {
            "$ref": "#/components/schemas/RecurringHolidayImpactCountsResponse"
          }
        },
        "required": [
          "combined",
          "current",
          "previous"
        ]
      },
      "RecurringHolidayMutationResponse": {
        "type": "object",
        "properties": {
          "holiday": {
            "$ref": "#/components/schemas/RecurringHolidayResponse"
          },
          "pendingConfigVersion": {
            "type": "integer",
            "format": "int64"
          },
          "impact": {
            "$ref": "#/components/schemas/RecurringHolidayImpactResponse"
          },
          "synchronization": {
            "$ref": "#/components/schemas/ScheduleSynchronizationResponse"
          }
        },
        "required": [
          "holiday",
          "impact",
          "pendingConfigVersion",
          "synchronization"
        ]
      },
      "RecurringHolidayResponse": {
        "type": "object",
        "properties": {
          "holidayId": {
            "type": "integer",
            "format": "int64"
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ]
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          },
          "reason": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "active",
          "dayOfWeek",
          "effectiveFrom",
          "holidayId",
          "reason",
          "version"
        ]
      },
...
      "ScheduleTemplatePreviewRequest": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "integer",
            "format": "int64"
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ]
          },
          "startTime": {
            "type": "string"
          }
        },
        "required": [
          "dayOfWeek",
          "startTime"
        ]
      },
...
      "RecurringHolidayPreviewRequest": {
        "type": "object",
        "properties": {
          "holidayId": {
            "type": "integer",
            "format": "int64"
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ]
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "dayOfWeek",
          "effectiveFrom"
        ]
      },
...
      "ScheduleActivationRequest": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          },
          "expectedConfigVersion": {
            "type": "integer",
            "format": "int64"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          }
        },
        "required": [
          "active",
          "expectedConfigVersion",
          "reason"
        ]
      },
...
      "ReservationCancellationPreviewResponse": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "integer",
            "format": "int64"
          },
          "timing": {
            "type": "string"
          },
          "responsibility": {
            "type": "string"
          },
          "couponAction": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "details",
          "fieldErrors",
          "message",
          "path",
          "status",
          "timestamp"
        ]
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}